JuliaUtils

Documentation for JuliaUtils.

JuliaUtils.JuliaUtils

JuliaUtils.aperiodic

JuliaUtils.cartesian_mesh

JuliaUtils.count_topological_charges

JuliaUtils.defect_vector_field

JuliaUtils.linspace

JuliaUtils.nearest_neighbour_distance

JuliaUtils.polar_mesh

JuliaUtils.polar_mesh

JuliaUtils.radial_distribution_function

JuliaUtils.JuliaUtils — Module

Main module for JuliaUtils.jl – my julia language utility functions.

Exports

aperiodic

cartesian_mesh

count_topological_charges

defect_vector_field

linspace

nearest_neighbour_distance

periodic_distance

polar_mesh

radial_distribution_function

read_data

read_xydata

@repeat

write_data

write_xydata

JuliaUtils.aperiodic — Method

Convert periodic data to an aperiodic one.

aperiodic(x)
aperiodic(x, p)

JuliaUtils.cartesian_mesh — Method

Mesh in cartesian coordinates.

cartesian_mesh(nx, lx; ny, ly, x0, y0)

JuliaUtils.count_topological_charges — Method

Count all topological zeros for a 2D orientation field.

count_topological_charges(θ; θ0, mx, my)

JuliaUtils.defect_vector_field — Method

Generate a topological defect vector field in two dimensions.

defect_vector_field(m, lx, nx; θ0, kwargs...)

JuliaUtils.linspace — Method

numpy like linspace.

linspace(x0, xn, n)

JuliaUtils.nearest_neighbour_distance — Method

Compute nearest neighbour distance for a given set of points in a 2D periodic box.

nearest_neighbour_distance(x, y, L)

JuliaUtils.polar_mesh — Method

Mesh in polar coordinates.

polar_mesh(nr, lr; nθ, lθ, r0, θ0)

JuliaUtils.polar_mesh — Method

Convert cartesian mesh to polar mesh.

polar_mesh(x, y)

JuliaUtils.radial_distribution_function — Method

Compute radial distribution function for a given set of points in a 2D periodic box.